Adaptation due to PR comments
