+++ src/main/java/io/github/bobocodebreskul/context/support/GeneralConstants.java
@@ -1,0 +1,13 @@
+package io.github.bobocodebreskul.context.support;
+
+import lombok.experimental.UtilityClass;
+
+/**
+ * Class to hold some general non specific constant values.
+ *
+ * @author Vitalii Katkov
+ */
+@UtilityClass
+public class GeneralConstants {
+    public static final String EMPTY = "";
+}
